modules/emow2v/requirements.txt
hcai-discover-utils
torch
torchaudio
librosa
transformers